move users to Bot local
this means possible contention but also all methods can access,
so it doesn't need to be passed around etc.
and if there are multiple loops for one bot, they share information.
an alternative would be to have a separate object to handle
these things and that could come later if it seems useful then