resequence dispatch/updates for .last/.user
here's the reasoning for the sequence.
we have to parse before we can even log to disk.
logging to disk is our highest priority, in case something
happens during command handling.

last message in general or for user is probably supposed to be
the last non-command, which means we have to dispatch before
we know whether to set conn.last_message and to update the user's
message. so updates are last.